phase 1

+++ src/components/CardNotes/CardNotes.jsx
@@ -8,6 +8,7 @@ import { Card } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
 import './cardNotes.css'
 import { CreateNotes } from '../ModalNotes/CreateNotes'
+import Swal from 'sweetalert2'
 
 export const CardNotes = ({ idNote, title, desc, category, isArchive = true, setRefresh }) => {
 
@@ -25,15 +26,34 @@ export const CardNotes = ({ idNote, title, desc, category, isArchive = true, set
     }
 
     const handleDelete = async () => {
-        await fetch(`https://apinotesql.herokuapp.com/api/notes/deleteNote/${idNote}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                await fetch(`https://apinotesql.herokuapp.com/api/notes/deleteNote/${idNote}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    }
+                }).then(res => res.json())
+                    .then(data => {
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                        setRefresh(true)
+                    })
             }
-        }).then(res => res.json())
-            .then(data => {
-                setRefresh(true)
-            })
+
+        })
+
     }
 
     const handleArchive = async () => {
@@ -45,12 +65,21 @@ export const CardNotes = ({ idNote, title, desc, category, isArchive = true, set
         })
         const response = await data.json();
         const { success, message } = response;
-        console.log(response)
         if (success) {
+            Swal.fire(
+                `${title} Archived!`,
+                `${message}`,
+                'success'
+            )
             setRefresh(true)
-        } else {
-
+        }else{
+            Swal.fire(
+                `${title} Archived!`,
+                `${message}`,
+                'error'
+            )
         }
+
     }
 
     const handleUnArchived = async () => {
@@ -63,10 +92,16 @@ export const CardNotes = ({ idNote, title, desc, category, isArchive = true, set
         const response = await data.json();
         const { success, message } = response;
         if (success) {
-            setRefresh(true)
-        } else {
-
+            Swal.fire(
+                'Unarchived!',
+                `${message}`,
+                'success'
+            )
+            setTimeout(() => {
+                location.reload();
+            }, 1000);
         }
+
     }
 
     return (
@@ -85,31 +120,6 @@ export const CardNotes = ({ idNote, title, desc, category, isArchive = true, set
                     <Button onClick={handleDelete}><AiOutlineDelete className='cart-icon' /></Button>
                 </Card.Body>
             </Card>
-            {/*  <div className="card">
-                <div className="tools">
-                    <div className="circle">
-                        <span className="red box"></span>
-                    </div>
-                    <div className="circle">
-                        <span className="yellow box"></span>
-                    </div>
-                    <div className="circle">
-                        <span className="green box"></span>
-                    </div>
-                </div>
-                <div className="card__content">
-                    <h2>title</h2>
-                    <p>{title}</p>
-                    <h2>Desciption</h2>
-                    <p>{desc}</p>
-                    {
-                        isArchive ? (<Button onClick={handleUnArchived}><RiInboxUnarchiveLine className='cart-icon' /></Button>) : (<Button onClick={handleArchive}><BiBox className='cart-icon' /></Button>)
-                    }
-                    <Button onClick={handleShow}><BsPencilSquare className='cart-icon' /></Button>
-                    <CreateNotes show={show} setShow={setShow} isEdit={isEdit} setRefresh={setRefresh} values={value} />
-                    <Button onClick={handleDelete}><AiOutlineDelete className='cart-icon' /></Button>
-                </div>
-            </div> */}
         </>
     )
 }
